Implement verify categorization subcommand
Add subcommand verify and it's first target property "categorization".
Calling verify on this property prints all tracks, that are saved but
not contained in a category playlist, and all tracks, that are in a
category playlist but not saved to the users libary. A category playlist
is a playlist whose name starts with "{A,B,C,D,E} - ".