Improve handling of 404 errors
The changes are necessary, because Spotify nowadays responses with
a lot of 404 errors. Previously in these cases the program would
abort. Now the given request os sent again, until a successful
response is returned.
At some places additional output is generated, which makes it easyer
to understand errors and to estimate the programs progress.